feat: add migrations and godoc comments
