+++ internal/pkg/auth/repo/postgres.go
@@ -7,38 +7,43 @@ import (
 	"errors"
 )
 
+// AuthRepo represents a repository for authentication.
 type AuthRepo struct {
 	db *sql.DB
 }
 
+// NewRepository creates a new instance of AuthRepo.
 func NewRepository(db *sql.DB) *AuthRepo {
 	return &AuthRepo{db: db}
 }
 
+// CreateUser creates a new user in the database.
 func (r *AuthRepo) CreateUser(ctx context.Context, user *models.User) error {
-	insert := `INSERT INTO users (id, login, phone, password_hash) VALUES ($1, $2, $3, $4)`
+	insert := `INSERT INTO users (id, login, password_hash) VALUES ($1, $2, $3, $4)`
 
-	if _, err := r.db.ExecContext(ctx, insert, user.ID, user.Login, user.Phone, user.PasswordHash); err != nil {
+	if _, err := r.db.ExecContext(ctx, insert, user.ID, user.Login, user.PasswordHash); err != nil {
 		return err
 	}
 	return nil
 }
 
+// GetUserByLogin retrieves a user from the database by their login.
 func (r *AuthRepo) GetUserByLogin(ctx context.Context, login string) (*models.User, error) {
-	query := `SELECT * FROM users WHERE login = $1`
+	query := `SELECT id, login FROM users WHERE login = $1`
 
 	res := r.db.QueryRowContext(ctx, query, login)
 
 	user := &models.User{
 		Login: login,
 	}
-	if err := res.Scan(&user.ID, &user.Login, &user.Phone, &user.PasswordHash); err != nil {
+	if err := res.Scan(&user.ID, &user.Login, &user.PasswordHash); err != nil {
 		return nil, err
 	}
 
 	return user, nil
 }
 
+// CheckUser checks if the user with the given login and password hash exists in the database.
 func (r *AuthRepo) CheckUser(ctx context.Context, login string, passwordHash string) (*models.User, error) {
 	user, err := r.GetUserByLogin(ctx, login)
 	if err != nil {
