+++ internal/pkg/utils/utils.go
@@ -7,11 +7,13 @@ import (
 	"net/http"
 )
 
+// WriteError writes an error response with the specified status code and message.
 func WriteError(w http.ResponseWriter, statusCode int, message string) {
 	w.WriteHeader(statusCode)
 	fmt.Fprintln(w, message)
 }
 
+// WriteResponse writes a JSON response with the specified status code and data.
 func WriteResponse(w http.ResponseWriter, statusCode int, response interface{}) error {
 	resp, err := json.Marshal(response)
 	if err != nil {
@@ -23,6 +25,7 @@ func WriteResponse(w http.ResponseWriter, statusCode int, response interface{}) 
 	return nil
 }
 
+// ReadRequestData reads and parses the request body into the provided structure.
 func ReadRequestData(r *http.Request, request interface{}) error {
 	data, err := io.ReadAll(r.Body)
 	if err != nil {
